requirements.txt
langchain==0.1.13
python-dotenv==1.0.1
streamlit
tiktoken==0.6.0 
faiss-cpu==1.7.2  
protobuf~=3.20,<5
google-generativeai 
langchain-community  
sentence-transformers==2.6.1  
InstructorEmbedding  
